+++ index.js
@@ -1,8 +1,10 @@
+import settings from './settings'
 import en from './en.js'
 import cs from './cs.js'
 import ja from './ja.js'
 
 export default {
+  settings,
   en,
   cs,
   ja